Tools: fix after RoomStruct's rework
