+++ Tools/data/scriptgen.cpp
@@ -15,7 +15,6 @@
 #include <iterator>
 #include <cctype>
 #include <cstring>
-#include "data/room_utils.h"
 
 namespace AGS
 {
@@ -242,37 +241,6 @@ String MakeVariablesScriptBody(std::vector<Variable> &vars)
     }
     body.Append("}\n");
     return body;
-}
-
-
-//-----------------------------------------------------------------------------
-// Room.ash
-//-----------------------------------------------------------------------------
-
-String MakeRoomScriptHeader(const RoomScNames &data)
-{
-    String header;
-    // Room Object names
-    for (const auto &obj : data.ObjectNames)
-    {
-        if (!obj.IsEmpty())
-        {
-            header.Append("import Object ");
-            header.Append(obj);
-            header.Append(";\n");
-        }
-    }
-    // Hotspot names
-    for (const auto &hot : data.HotspotNames)
-    {
-        if (!hot.IsEmpty())
-        {
-            header.Append("import Hotspot ");
-            header.Append(hot);
-            header.Append(";\n");
-        }
-    }
-    return header;
 }
 
 } // namespace DataUtil
